+++ R/simulation-spades.R
@@ -741,6 +741,7 @@ setMethod(
                         notOlderThan,
                         ...) {
 
+    .pkgEnv$.sim <- NULL # Clear anything that was here.
     # set the options("spades.xxxPath") to the values in the sim@paths
     oldGetPaths <- getPaths()
     do.call(setPaths, append(sim@paths, list(silent = TRUE)))
@@ -760,13 +761,14 @@ setMethod(
       if (!.pkgEnv[["skipNamespacing"]])
         .modifySearchPath(.pkgEnv$searchPath, removeOthers = TRUE)
       rm(".timeunits", envir = sim@.xData)
-      if (isTRUE(getOption("spades.saveSimOnExit", FALSE)))
-        if (isTRUE(.pkgEnv$.cleanEnd)) {
-          rm(list = ".sim", envir = .pkgEnv)
-        } else{
-          message("Because of a failed spades call, the sim object at the time of failure ",
-                  "was saved in SpaDES.core:::.pkgEnv$.sim . It will be deleted on next call to spades")
+      if (isTRUE(getOption("spades.saveSimOnExit", FALSE))) {
+        if (!isTRUE(.pkgEnv$.cleanEnd)) {
+          .pkgEnv$.sim <- sim
+          message(crayon::magenta("Because of a failed spades call, the sim object at the time of failure ",
+                  "was saved in \nSpaDES.core:::.pkgEnv$.sim . It will be deleted on next call to spades"))
         }
+        .pkgEnv$.cleanEnd <- NULL
+      }
     }, add = TRUE)
 
     if (!is.null(.plotInitialTime)) {
@@ -825,13 +827,8 @@ setMethod(
 
     sim@.xData[["._firstEventClockTime"]] <- Sys.time()
 
-    saveSimOnExit <- getOption("spades.saveSimOnExit", FALSE)
     while (sim@simtimes[["current"]] <= sim@simtimes[["end"]]) {
       sim <- doEvent(sim, debug = debug, notOlderThan = notOlderThan)  # process the next event
-      if (saveSimOnExit) {
-        .pkgEnv$.sim <- sim
-      }
-
 
       # Conditional Scheduling -- adds only 900 nanoseconds per event, if none exist
       if (exists("._conditionalEvents", envir = sim, inherits = FALSE)) {